pyproject.toml
[project]
name = "handwritten-digit-recogniser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy==1.25.2",
    "pillow==10.0.0",
    "torch==2.0.1",
    "torchvision==0.15.2",
]
